refactor: update jobeditor namings
