+++ lib/lightning_web/live/job_live/job_builder_components.ex
@@ -11,13 +11,13 @@ defmodule LightningWeb.JobLive.JobBuilderComponents do
   attr :source, :string, required: true
   attr :rest, :global
 
-  jsx("assets/js/job-editor/JobEditorMain.tsx")
+  jsx("assets/js/job-editor/JobEditor.tsx")
 
   def job_editor_component(assigns) do
     assigns = assigns |> assign(disabled: assigns.disabled |> to_string())
 
     ~H"""
-    <.JobEditorMain
+    <.JobEditor
       job_id={@job_id}
       adaptor={@adaptor}
       source={@source}
